feat : 본인 검증이 관리자는 패스하도록 변경, post의 좋아요 클릭/언클릭, 특정유저 게시물 모아보기, 특정유저 채팅…

+++ project_linkus_back/src/main/java/com/my/project_linkus_back/posts/repository/PostLikesRepository.java
@@ -11,5 +11,5 @@ public interface PostLikesRepository extends JpaRepository<PostLikes, Long> {
 
     void deleteByPostAndUser(Posts post, Users user);
 
-    long countByPost(Posts post);
+    int countByPost(Posts post);
 }
